backend/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
python-dotenv==1.0.1
langchain==0.1.9
langchain-openai==0.0.5
qdrant-client==1.7.3
pydantic==2.6.1 